Datadynamics_Assignment3_Track1/Module_B/callhub_backend/requirements.txt
Flask
mysql-connector-python
dotenv
Faker
requests
matplotlib
flask_mysqldb
bcrypt
locust
